+++ src/index.ts
@@ -3,7 +3,10 @@
 
 export type { AppProps } from './App';
 export { default as App } from './App';
-export { default as createFontStack } from './util/createFontStack';
-export { default as theme } from './util/theme';
+export type { ButtonProps } from './Button';
+export { default as Button } from './Button';
 export type { ThemeProps } from './Theme';
 export { default as Theme } from './Theme';
+
+export { default as createFontStack } from './util/createFontStack';
+export { default as theme } from './util/theme';